requirements.txt
streamlit
pandas
scikit-learn
openpyxl
python-docx
python-pptx
PyPDF2
tika
joblib
